chore(odyssey-react-mui): abstracting more logic to hook
